댓글 예외처리 완료

+++ routes/commentPage.js
@@ -11,14 +11,17 @@ router.post("/create", authMiddleWare, async (req, res) => {
     const { userInfo } = res.locals;
     const { userId, userName } = userInfo; // userName은 User 테이블에서 참조해올것
 
-    await Comment.create({
-      postId,
-      comment,
-      userId,
-      userName,
-    });
-
-    res.status(200).json({ msg: "comment create complete." });
+    if (!comment) {
+      return res.status(400).json({ msg: "빈칸 없이 입력해주세요" });
+    } else {
+      await Comment.create({
+        postId,
+        comment,
+        userId,
+        userName,
+      });
+      res.status(200).json({ msg: "comment create complete." });
+    }
   } catch (error) {
     console.log(error);
     console.log("commentPage.js --> 댓글 등록에서 에러남");
@@ -43,11 +46,15 @@ router.get("/updateRawData", authMiddleWare, async (req, res) => {
       where: { postId, commentId },
     });
 
-    // 카카오, 구글에서 제공한 userId와 postId로 DB에서 꺼내온 userId가 같은지 비교
-    if (commentList.userId === userId) {
-      res.status(200).json({ commentList });
+    if (commentList) {
+      // 카카오, 구글에서 제공한 userId와 postId로 DB에서 꺼내온 userId가 같은지 비교
+      if (commentList.userId === userId) {
+        res.status(200).json({ commentList });
+      } else {
+        res.status(403).json({ msg: "본인의 댓글이 아닙니다." });
+      }
     } else {
-      res.status(403).json({ msg: "본인의 댓글이 아닙니다." });
+      res.status(404).json({ msg: "해당 댓글이 존재하지 않습니다." });
     }
   } catch (error) {
     console.log(error);
@@ -66,23 +73,30 @@ router.patch("/update", authMiddleWare, async (req, res) => {
     const { userInfo } = res.locals;
     const { userId, userName } = userInfo;
 
+    if (!comment) {
+      return res.status(400).json({ msg: "빈칸 없이 입력해주세요" });
+    }
+
     let commentList = await Comment.findOne({
       logging: false,
       attributes: ["userId"],
       where: { commentId },
     });
-
-    // 카카오, 구글에서 제공한 userId와 postId로 DB에서 꺼내온 userId가 같은지 비교
-    if (commentList.userId === userId) {
-      await Comment.update(
-        { comment },
-        {
-          where: { postId, commentId },
-        }
-      );
-      res.status(200).json({ msg: "comment update complete." });
+    if (commentList) {
+      // 카카오, 구글에서 제공한 userId와 postId로 DB에서 꺼내온 userId가 같은지 비교
+      if (commentList.userId === userId) {
+        await Comment.update(
+          { comment },
+          {
+            where: { postId, commentId },
+          }
+        );
+        res.status(200).json({ msg: "comment update complete." });
+      } else {
+        res.status(403).json({ msg: "본인의 댓글이 아닙니다." });
+      }
     } else {
-      res.status(403).json({ msg: "본인의 댓글이 아닙니다." });
+      res.status(404).json({ msg: "해당 댓글이 존재하지 않습니다." });
     }
   } catch (error) {
     console.log(error);
@@ -105,14 +119,18 @@ router.delete("/delete", authMiddleWare, async (req, res) => {
       where: { commentId },
     });
 
-    // 카카오, 구글에서 제공한 userId와 postId로 DB에서 꺼내온 userId가 같은지 비교
-    if (commentList.userId === userId) {
-      await Comment.destroy({
-        where: { commentId },
-      });
-      res.status(200).json({ msg: "comment delete complete." });
+    if (commentList) {
+      // 카카오, 구글에서 제공한 userId와 postId로 DB에서 꺼내온 userId가 같은지 비교
+      if (commentList.userId === userId) {
+        await Comment.destroy({
+          where: { commentId },
+        });
+        res.status(200).json({ msg: "comment delete complete." });
+      } else {
+        res.status(403).json({ msg: "본인의 댓글이 아닙니다." });
+      }
     } else {
-      res.status(403).json({ msg: "본인의 댓글이 아닙니다." });
+      res.status(404).json({ msg: "해당 댓글이 존재하지 않습니다." });
     }
   } catch (error) {
     console.log(error);
